w64app-main/requirements.txt
psycopg2
sqlalchemy
pandas
numpy
jupyter
python-dotenv
matplotlib
prophet